Move frontend test files to __tests__ folder
