+++ frontend/src/components/__tests__/App.test.jsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom'
-import App from './App'
+import App from '../App'
 
 test('App should render App fragment component correctly', async () => {
   render(<App />)
