+++ frontend/src/components/__tests__/Home.test.jsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom'
-import Home from './Home'
+import Home from '../Home'
 
 test('Home should render Home fragment component correctly', async () => {
   render(<Home />)
