+++ frontend/src/components/__tests__/Navbar.test.jsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom'
-import Navbar from './Navbar'
+import Navbar from '../Navbar'
 
 test('Navbar should render Navbar component correctly', async () => {
   render(<Navbar />)
