+++ frontend/src/components/__tests__/Register.test.jsx
@@ -1,11 +1,12 @@
 import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom'
-import Login from './Login'
+import Register from '../Register'
 
-test('Login should render Login form component correctly', async () => {
-  render(<Login />)
+test('Register form should render Register form component correctly', async () => {
+  render(<Register />)
 
   expect(await screen.getByTestId('test-email')).toBeInTheDocument()
+  expect(await screen.getByTestId('test-username')).toBeInTheDocument()
   expect(await screen.getByTestId('test-password')).toBeInTheDocument()
   expect(await screen.getByTestId('test-submit-button')).toBeInTheDocument()
 })
